* use google collection class Lists * fix NPE if no selenium tests exists

+++ src/play/modules/webdrive/SeleniumTests2JUnitXML.java
@@ -10,17 +10,17 @@ import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.stream.StreamResult;
 import javax.xml.transform.stream.StreamSource;
 
-import org.testng.collections.Lists;
-
 import play.Logger;
 import play.libs.IO;
+
+import com.google.common.collect.Lists;
 
 public class SeleniumTests2JUnitXML {
   private static final String REGEX_HTML_META_TAG_CLOSED = "<meta $1 />";
   private static final String REGEX_HTML_META_TAG_INVALID = "<meta(.*)>";
   private static final String REGEX_HTML_ENTITY_WHITESPACE = "&nbsp;";
   private static final String REGEX_HTML_WHITESPACE = " ";
-  
+
   private static final String XML_EXT = ".xml";
   private static final String HTML_EXT = ".html";
   static final String XALAN_INDENT_AMOUNT = "{http://xml.apache.org/xslt}" + "indent-amount";
@@ -35,31 +35,37 @@ public class SeleniumTests2JUnitXML {
     Logger.debug("   translation: plays selenium test (html result page) --> junit xml file (bamboo readable)");
     for (String path : paths) {
       Logger.debug("process " + path);
-      List<File> seleniumTests = Arrays.asList(new File(path).listFiles());
-      for (File seleniumTest : seleniumTests) {
-        try {
-          TransformerFactory tFactory = TransformerFactory.newInstance();
-          Transformer transformer = tFactory.newTransformer(new StreamSource(xslFile));
-          transformer.setOutputProperty(XALAN_INDENT_AMOUNT, "4");
-          if (seleniumTest != null) {
-            String absolutePathSeleniumTest = seleniumTest.getAbsolutePath();
-            if (absolutePathSeleniumTest.endsWith(HTML_EXT)) {
+      File seleniumTestFolder = new File(path);
+      if (seleniumTestFolder.exists() || seleniumTestFolder.isDirectory()) {
+        List<File> seleniumTests = Arrays.asList(seleniumTestFolder.listFiles());
+        for (File seleniumTest : seleniumTests) {
+          try {
+            TransformerFactory tFactory = TransformerFactory.newInstance();
+            Transformer transformer = tFactory.newTransformer(new StreamSource(xslFile));
+            transformer.setOutputProperty(XALAN_INDENT_AMOUNT, "4");
+            if (seleniumTest != null) {
+              String absolutePathSeleniumTest = seleniumTest.getAbsolutePath();
+              if (absolutePathSeleniumTest.endsWith(HTML_EXT)) {
 
-              File seleniumTestFile = new File(absolutePathSeleniumTest);
-              String htmlResultContent = IO.readContentAsString(seleniumTestFile);
-              htmlResultContent = htmlResultContent.replaceAll(REGEX_HTML_META_TAG_INVALID, REGEX_HTML_META_TAG_CLOSED);
-              htmlResultContent = htmlResultContent.replaceAll(REGEX_HTML_ENTITY_WHITESPACE, REGEX_HTML_WHITESPACE);
-              IO.writeContent(htmlResultContent, seleniumTestFile);
-              File resultFile = new File(seleniumTestFile + XML_EXT);
-              FileOutputStream resultOutputStream = new FileOutputStream(resultFile);
-              StreamSource streamSource = new StreamSource(seleniumTest);
-              StreamResult streamResultAfterXSLProcessing = new StreamResult(resultOutputStream);
-              transformer.transform(streamSource, streamResultAfterXSLProcessing);
+                File seleniumTestFile = new File(absolutePathSeleniumTest);
+                String htmlResultContent = IO.readContentAsString(seleniumTestFile);
+                htmlResultContent = htmlResultContent.replaceAll(REGEX_HTML_META_TAG_INVALID, REGEX_HTML_META_TAG_CLOSED);
+                htmlResultContent = htmlResultContent.replaceAll(REGEX_HTML_ENTITY_WHITESPACE, REGEX_HTML_WHITESPACE);
+                IO.writeContent(htmlResultContent, seleniumTestFile);
+                File resultFile = new File(seleniumTestFile + XML_EXT);
+                FileOutputStream resultOutputStream = new FileOutputStream(resultFile);
+                StreamSource streamSource = new StreamSource(seleniumTest);
+                StreamResult streamResultAfterXSLProcessing = new StreamResult(resultOutputStream);
+                transformer.transform(streamSource, streamResultAfterXSLProcessing);
+              }
             }
+          } catch (Throwable t) {
+            Logger.error(t, t.getMessage());
           }
-        } catch (Throwable t) {
-          Logger.error(t, t.getMessage());
         }
+      } else {
+        Logger.error("path '%s' is not ok: exists: '%s', is directory: '%s'", path, Boolean.toString(seleniumTestFolder.exists()),
+            Boolean.toString(seleniumTestFolder.isDirectory()));
       }
     }
   }
